+++ dask/dataframe/methods.py
@@ -5,7 +5,7 @@ import pandas as pd
 from pandas.types.concat import union_categoricals
 from toolz import partition
 
-from .utils import PANDAS_VERSION, is_categorical_dtype
+from .utils import is_categorical_dtype
 
 
 # ---------------------------------
@@ -201,88 +201,58 @@ def concat(dfs, axis=0, join='outer', uniform=False):
         necessarily categories). Default is False.
     """
     if axis == 1:
-        if PANDAS_VERSION >= '0.18.1':
-            return pd.concat(dfs, axis=axis, join=join)
-
-        # Concat with empty Series with axis=1 will not affect to the
-        # result. Special handling is needed in each partition
-        # Can be removed after pandas 0.18.1 or later
-        # see https://github.com/pandas-dev/pandas/pull/12846
-        if join == 'outer':
-            # outer concat should keep all empty Series
-
-            # input must include one non-empty data at least
-            # because of the alignment
-            first = [df for df in dfs if len(df) > 0][0]
-
-            def _pad(base, fillby):
-                if isinstance(base, pd.Series) and len(base) == 0:
-                    # use aligned index to keep index for outer concat
-                    return pd.Series([np.nan] * len(fillby),
-                                     index=fillby.index, name=base.name)
-                else:
-                    return base
-
-            dfs = [_pad(df, first) for df in dfs]
-        else:
-            # inner concat should result in empty if any input is empty
-            if any(len(df) == 0 for df in dfs):
-                dfs = [pd.DataFrame(columns=df.columns)
-                       if isinstance(df, pd.DataFrame) else
-                       pd.Series(name=df.name) for df in dfs]
-
         return pd.concat(dfs, axis=axis, join=join)
+
+    # Support concatenating indices along axis 0
+    if isinstance(dfs[0], pd.Index):
+        if isinstance(dfs[0], pd.CategoricalIndex):
+            return pd.CategoricalIndex(union_categoricals(dfs),
+                                       name=dfs[0].name)
+        return dfs[0].append(dfs[1:])
+
+    # Handle categorical index separately
+    if isinstance(dfs[0].index, pd.CategoricalIndex):
+        dfs2 = [df.reset_index(drop=True) for df in dfs]
+        ind = concat([df.index for df in dfs])
     else:
-        # Support concatenating indices along axis 0
-        if isinstance(dfs[0], pd.Index):
-            if isinstance(dfs[0], pd.CategoricalIndex):
-                return pd.CategoricalIndex(union_categoricals(dfs),
-                                           name=dfs[0].name)
-            return dfs[0].append(dfs[1:])
+        dfs2 = dfs
+        ind = None
 
-        # Handle categorical index separately
-        if isinstance(dfs[0].index, pd.CategoricalIndex):
-            dfs2 = [df.reset_index(drop=True) for df in dfs]
-            ind = concat([df.index for df in dfs])
+    # Concatenate the partitions together, handling categories as needed
+    if (isinstance(dfs2[0], pd.DataFrame) if uniform else
+            any(isinstance(df, pd.DataFrame) for df in dfs2)):
+        if uniform:
+            dfs3 = dfs2
+            cat_mask = dfs2[0].dtypes == 'category'
         else:
-            dfs2 = dfs
-            ind = None
+            # When concatenating mixed dataframes and series on axis 1, Pandas
+            # converts series to dataframes with a single column named 0, then
+            # concatenates.
+            dfs3 = [df if isinstance(df, pd.DataFrame) else
+                    df.rename(0).to_frame() for df in dfs2]
+            cat_mask = pd.concat([(df.dtypes == 'category').to_frame().T
+                                  for df in dfs3], join=join).any()
 
-        # Concatenate the partitions together, handling categories as needed
-        if (isinstance(dfs2[0], pd.DataFrame) if uniform else
-                any(isinstance(df, pd.DataFrame) for df in dfs2)):
-            if uniform:
-                dfs3 = dfs2
-                cat_mask = dfs2[0].dtypes == 'category'
-            else:
-                # When concatenating mixed dataframes and series on axis 1,
-                # Pandas converts series to dataframes with a single column
-                # named 0, then concatenates.
-                dfs3 = [df if isinstance(df, pd.DataFrame) else
-                        df.rename(0).to_frame() for df in dfs2]
-                cat_mask = pd.concat([(df.dtypes == 'category').to_frame().T
-                                      for df in dfs3], join=join).any()
-
-            if cat_mask.any():
-                not_cat = cat_mask[~cat_mask].index
-                out = pd.concat([df[df.columns.intersection(not_cat)]
-                                 for df in dfs3], join=join)
-                for col in cat_mask.index.difference(not_cat):
-                    out[col] = union_categoricals([df.get(col) for df in dfs3])
-                out = out.reindex_axis(cat_mask.index, axis=1)
-            else:
-                out = pd.concat(dfs3, join=join)
+        if cat_mask.any():
+            not_cat = cat_mask[~cat_mask].index
+            out = pd.concat([df[df.columns.intersection(not_cat)]
+                             for df in dfs3], join=join)
+            for col in cat_mask.index.difference(not_cat):
+                out[col] = union_categoricals([df.get(col) for df in dfs3])
+            out = out.reindex_axis(cat_mask.index, axis=1)
         else:
-            if is_categorical_dtype(dfs2[0].dtype):
-                if ind is None:
-                    ind = concat([df.index for df in dfs2])
-                return pd.Series(union_categoricals(dfs2), index=ind,
-                                 name=dfs2[0].name)
-            out = pd.concat(dfs2, join=join)
-        # Re-add the index if needed
-        if ind is not None:
-            out.index = ind
-        return out
+            out = pd.concat(dfs3, join=join)
+    else:
+        if is_categorical_dtype(dfs2[0].dtype):
+            if ind is None:
+                ind = concat([df.index for df in dfs2])
+            return pd.Series(union_categoricals(dfs2), index=ind,
+                             name=dfs2[0].name)
+        out = pd.concat(dfs2, join=join)
+    # Re-add the index if needed
+    if ind is not None:
+        out.index = ind
+    return out
 
 
 def merge(left, right, how, left_on, right_on,
