+++ dask/dataframe/tests/test_dataframe.py
@@ -14,8 +14,7 @@ from dask import delayed
 from dask.utils import ignoring, put_lines
 import dask.dataframe as dd
 
-from dask.dataframe.core import (repartition_divisions, aca, _concat,
-                                 _Frame, Scalar)
+from dask.dataframe.core import repartition_divisions, aca, _concat, Scalar
 from dask.dataframe.methods import boundary_slice
 from dask.dataframe.utils import assert_eq, make_meta, assert_max_deps
 
@@ -955,171 +954,6 @@ def test_unknown_divisions():
 
     assert_eq(d.a.sum(), full.a.sum())
     assert_eq(d.a + d.b + 1, full.a + full.b + 1)
-
-
-def test_concat2():
-    dsk = {('x', 0): pd.DataFrame({'a': [1, 2, 3], 'b': [4, 5, 6]}),
-           ('x', 1): pd.DataFrame({'a': [4, 5, 6], 'b': [3, 2, 1]}),
-           ('x', 2): pd.DataFrame({'a': [7, 8, 9], 'b': [0, 0, 0]})}
-    meta = make_meta({'a': 'i8', 'b': 'i8'})
-    a = dd.DataFrame(dsk, 'x', meta, [None, None])
-    dsk = {('y', 0): pd.DataFrame({'a': [10, 20, 30], 'b': [40, 50, 60]}),
-           ('y', 1): pd.DataFrame({'a': [40, 50, 60], 'b': [30, 20, 10]}),
-           ('y', 2): pd.DataFrame({'a': [70, 80, 90], 'b': [0, 0, 0]})}
-    b = dd.DataFrame(dsk, 'y', meta, [None, None])
-
-    dsk = {('y', 0): pd.DataFrame({'b': [10, 20, 30], 'c': [40, 50, 60]}),
-           ('y', 1): pd.DataFrame({'b': [40, 50, 60], 'c': [30, 20, 10]})}
-    meta = make_meta({'b': 'i8', 'c': 'i8'})
-    c = dd.DataFrame(dsk, 'y', meta, [None, None])
-
-    dsk = {('y', 0): pd.DataFrame({'b': [10, 20, 30], 'c': [40, 50, 60],
-                                   'd': [70, 80, 90]}),
-           ('y', 1): pd.DataFrame({'b': [40, 50, 60], 'c': [30, 20, 10],
-                                   'd': [90, 80, 70]},
-                                  index=[3, 4, 5])}
-    meta = make_meta({'b': 'i8', 'c': 'i8', 'd': 'i8'},
-                     index=pd.Index([], 'i8'))
-    d = dd.DataFrame(dsk, 'y', meta, [0, 3, 5])
-
-    cases = [[a, b], [a, c], [a, d]]
-    assert dd.concat([a]) is a
-    for case in cases:
-        result = dd.concat(case)
-        pdcase = [_c.compute() for _c in case]
-
-        assert result.npartitions == case[0].npartitions + case[1].npartitions
-        assert result.divisions == (None, ) * (result.npartitions + 1)
-        assert_eq(pd.concat(pdcase), result)
-        assert result.dask == dd.concat(case).dask
-
-        result = dd.concat(case, join='inner')
-        assert result.npartitions == case[0].npartitions + case[1].npartitions
-        assert result.divisions == (None, ) * (result.npartitions + 1)
-        assert_eq(pd.concat(pdcase, join='inner'), result)
-        assert result.dask == dd.concat(case, join='inner').dask
-
-
-def test_concat3():
-    pdf1 = pd.DataFrame(np.random.randn(6, 5),
-                        columns=list('ABCDE'), index=list('abcdef'))
-    pdf2 = pd.DataFrame(np.random.randn(6, 5),
-                        columns=list('ABCFG'), index=list('ghijkl'))
-    pdf3 = pd.DataFrame(np.random.randn(6, 5),
-                        columns=list('ABCHI'), index=list('mnopqr'))
-    ddf1 = dd.from_pandas(pdf1, 2)
-    ddf2 = dd.from_pandas(pdf2, 3)
-    ddf3 = dd.from_pandas(pdf3, 2)
-
-    result = dd.concat([ddf1, ddf2])
-    assert result.divisions == ddf1.divisions[:-1] + ddf2.divisions
-    assert result.npartitions == ddf1.npartitions + ddf2.npartitions
-    assert_eq(result, pd.concat([pdf1, pdf2]))
-
-    assert_eq(dd.concat([ddf1, ddf2], interleave_partitions=True),
-              pd.concat([pdf1, pdf2]))
-
-    result = dd.concat([ddf1, ddf2, ddf3])
-    assert result.divisions == (ddf1.divisions[:-1] + ddf2.divisions[:-1] +
-                                ddf3.divisions)
-    assert result.npartitions == (ddf1.npartitions + ddf2.npartitions +
-                                  ddf3.npartitions)
-    assert_eq(result, pd.concat([pdf1, pdf2, pdf3]))
-
-    assert_eq(dd.concat([ddf1, ddf2, ddf3], interleave_partitions=True),
-              pd.concat([pdf1, pdf2, pdf3]))
-
-
-def test_concat4_interleave_partitions():
-    pdf1 = pd.DataFrame(np.random.randn(10, 5),
-                        columns=list('ABCDE'), index=list('abcdefghij'))
-    pdf2 = pd.DataFrame(np.random.randn(13, 5),
-                        columns=list('ABCDE'), index=list('fghijklmnopqr'))
-    pdf3 = pd.DataFrame(np.random.randn(13, 6),
-                        columns=list('CDEXYZ'), index=list('fghijklmnopqr'))
-
-    ddf1 = dd.from_pandas(pdf1, 2)
-    ddf2 = dd.from_pandas(pdf2, 3)
-    ddf3 = dd.from_pandas(pdf3, 2)
-
-    msg = ('All inputs have known divisions which cannot be '
-           'concatenated in order. Specify '
-           'interleave_partitions=True to ignore order')
-
-    cases = [[ddf1, ddf1], [ddf1, ddf2], [ddf1, ddf3], [ddf2, ddf1],
-             [ddf2, ddf3], [ddf3, ddf1], [ddf3, ddf2]]
-    for case in cases:
-        pdcase = [c.compute() for c in case]
-
-        with tm.assertRaisesRegexp(ValueError, msg):
-            dd.concat(case)
-
-        assert_eq(dd.concat(case, interleave_partitions=True),
-                  pd.concat(pdcase))
-        assert_eq(dd.concat(case, join='inner', interleave_partitions=True),
-                  pd.concat(pdcase, join='inner'))
-
-    msg = "'join' must be 'inner' or 'outer'"
-    with tm.assertRaisesRegexp(ValueError, msg):
-        dd.concat([ddf1, ddf1], join='invalid', interleave_partitions=True)
-
-
-def test_concat5():
-    pdf1 = pd.DataFrame(np.random.randn(7, 5),
-                        columns=list('ABCDE'), index=list('abcdefg'))
-    pdf2 = pd.DataFrame(np.random.randn(7, 6),
-                        columns=list('FGHIJK'), index=list('abcdefg'))
-    pdf3 = pd.DataFrame(np.random.randn(7, 6),
-                        columns=list('FGHIJK'), index=list('cdefghi'))
-    pdf4 = pd.DataFrame(np.random.randn(7, 5),
-                        columns=list('FGHAB'), index=list('cdefghi'))
-    pdf5 = pd.DataFrame(np.random.randn(7, 5),
-                        columns=list('FGHAB'), index=list('fklmnop'))
-
-    ddf1 = dd.from_pandas(pdf1, 2)
-    ddf2 = dd.from_pandas(pdf2, 3)
-    ddf3 = dd.from_pandas(pdf3, 2)
-    ddf4 = dd.from_pandas(pdf4, 2)
-    ddf5 = dd.from_pandas(pdf5, 3)
-
-    cases = [[ddf1, ddf2], [ddf1, ddf3], [ddf1, ddf4], [ddf1, ddf5],
-             [ddf3, ddf4], [ddf3, ddf5], [ddf5, ddf1, ddf4], [ddf5, ddf3],
-             [ddf1.A, ddf4.A], [ddf2.F, ddf3.F], [ddf4.A, ddf5.A],
-             [ddf1.A, ddf4.F], [ddf2.F, ddf3.H], [ddf4.A, ddf5.B],
-             [ddf1, ddf4.A], [ddf3.F, ddf2], [ddf5, ddf1.A, ddf2]]
-
-    for case in cases:
-        pdcase = [c.compute() for c in case]
-
-        assert_eq(dd.concat(case, interleave_partitions=True),
-                  pd.concat(pdcase))
-
-        assert_eq(dd.concat(case, join='inner', interleave_partitions=True),
-                  pd.concat(pdcase, join='inner'))
-
-        assert_eq(dd.concat(case, axis=1), pd.concat(pdcase, axis=1))
-
-        assert_eq(dd.concat(case, axis=1, join='inner'),
-                  pd.concat(pdcase, axis=1, join='inner'))
-
-    # Dask + pandas
-    cases = [[ddf1, pdf2], [ddf1, pdf3], [pdf1, ddf4],
-             [pdf1.A, ddf4.A], [ddf2.F, pdf3.F],
-             [ddf1, pdf4.A], [ddf3.F, pdf2], [ddf2, pdf1, ddf3.F]]
-
-    for case in cases:
-        pdcase = [c.compute() if isinstance(c, _Frame) else c for c in case]
-
-        assert_eq(dd.concat(case, interleave_partitions=True),
-                  pd.concat(pdcase))
-
-        assert_eq(dd.concat(case, join='inner', interleave_partitions=True),
-                  pd.concat(pdcase, join='inner'))
-
-        assert_eq(dd.concat(case, axis=1), pd.concat(pdcase, axis=1))
-
-        assert_eq(dd.concat(case, axis=1, join='inner'),
-                  pd.concat(pdcase, axis=1, join='inner'))
 
 
 def test_append():
